tests: consolidate validation unit tests into test_validate.py
test_integration_schema.py didn't contain integration tests...
